+++ src/cosmospy/_transaction.py
@@ -3,19 +3,16 @@ from __future__ import annotations
 import base64
 import hashlib
 import json
-from typing import Any
 
 import ecdsa
 
-from cosmospy._wallet import DEFAULT_BECH32_HRP, privkey_to_address, privkey_to_pubkey
-from cosmospy.typing import SyncModeApi, SyncModeRpc
-
-from cosmospy.generated import  bank_tx_pb2 as bank_msg
-from cosmospy.generated import coin_pb2 as coin
-from cosmospy.generated import tx_pb2 as tx
-from cosmospy.generated import keys_pb2 as keys
-from cosmospy.generated import any_pb2 as any
-
+from ._wallet import DEFAULT_BECH32_HRP, privkey_to_address, privkey_to_pubkey
+from .generated import any_pb2 as any
+from .generated import bank_tx_pb2 as bank_msg
+from .generated import coin_pb2 as coin
+from .generated import keys_pb2 as keys
+from .generated import tx_pb2 as tx
+from .typing import SyncModeApi, SyncModeRpc
 
 
 class Transaction:
@@ -51,22 +48,22 @@ class Transaction:
         self._fee_denom = fee_denom
         self._gas = gas
         self._memo = memo
+        self._tx_body.memo = memo
         self._chain_id = chain_id
         self._hrp = hrp
         self._sync_mode_rpc = sync_mode_rpc
         self._sync_mode_api = sync_mode_api
-
 
     def add_transfer(self, recipient: str, amount: int, denom: str = "uatom") -> None:
         msg = bank_msg.MsgSend()
         msg.from_address = privkey_to_address(self._privkey, hrp=self._hrp)
         msg.to_address = recipient
 
-        amount = coin.Coin()
-        amount.denom = "6"
-        amount.amount = str(amount)
+        coin_amount = coin.Coin()
+        coin_amount.denom = denom
+        coin_amount.amount = str(amount)
 
-        msg.amount.append(amount)
+        msg.amount.append(coin_amount)
 
         msg_any = any.Any()
         msg_any.Pack(msg)
@@ -74,30 +71,19 @@ class Transaction:
         self._tx_body.messages.append(msg_any)
 
     def get_pushable_rpc(self) -> str:
-        self._tx_raw.body_bytes = self._tx_body.SerializeToString()
-        self._tx_raw.auth_info_bytes = self._get_auth_info().SerializeToString()
-        self._tx_raw.signatures.append(self._get_signatures())
-        raw_tx = self._tx_raw.SerializeToString()
-        tx_bytes = bytes(raw_tx)
-        tx_b64 = base64.b64encode(tx_bytes).decode("utf-8")
+        tx_b64 = self._get_b64_encoded()
         return json.dumps(
             {"jsonrpc": "2.0", "id": 1, "method": self._sync_mode_rpc, "params": {"tx": tx_b64}}
         )
 
     def get_pushable_api(self) -> str:
-        self._tx_raw.body_bytes = self._tx_body.SerializeToString()
-        self._tx_raw.auth_info_bytes = self._get_auth_info().SerializeToString()
-        self._tx_raw.signatures.append(self._get_signatures())
-        raw_tx = self._tx_raw.SerializeToString()
-        tx_bytes = bytes(raw_tx)
-        tx_b64 = base64.b64encode(tx_bytes).decode("utf-8")
+        tx_b64 = self._get_b64_encoded()
         return json.dumps(
             {
                 "tx_bytes": tx_b64,
                 "mode": self._sync_mode_rpc
             }
         )
-
 
     def _get_signatures(self):
         privkey = ecdsa.SigningKey.from_string(self._privkey, curve=ecdsa.SECP256k1)
@@ -143,3 +129,9 @@ class Transaction:
         signer_infos.mode_info.single.mode = 1
         return signer_infos
 
+    def _get_b64_encoded(self):
+        self._tx_raw.body_bytes = self._tx_body.SerializeToString()
+        self._tx_raw.auth_info_bytes = self._get_auth_info().SerializeToString()
+        self._tx_raw.signatures.append(self._get_signatures())
+        raw_tx = self._tx_raw.SerializeToString()
+        return base64.b64encode(bytes(raw_tx)).decode("utf-8")