+++ src/cosmospy/_wallet.py
@@ -5,8 +5,8 @@ import ecdsa
 import hdwallets
 import mnemonic
 
-from cosmospy import BIP32DerivationError
-from cosmospy.typing import Wallet
+from . import BIP32DerivationError
+from .typing import Wallet
 
 DEFAULT_DERIVATION_PATH = "m/44'/118'/0'/0/0"
 DEFAULT_BECH32_HRP = "cosmos"